Use int64_t

+++ wpiutil/src/main/native/include/wpi/util/Color.hpp
@@ -25,10 +25,9 @@ constexpr double ceil_int(double x, double x_whole) noexcept {
 
 constexpr double ceil(double x) noexcept {
   if (std::is_constant_evaluated()) {
-    return (
-        (x < 0.0 ? -x : x) >= 4503599627370496.
-            ? x
-            : ceil_int(x, static_cast<double>(static_cast<long long int>(x))));
+    return ((x < 0.0 ? -x : x) >= 4503599627370496.
+                ? x
+                : ceil_int(x, static_cast<double>(static_cast<int64_t>(x))));
   } else {
     return std::ceil(x);
   }
